Reuse list selection menu module
