+++ codex-rs/tui/src/bottom_pane/file_search_popup.rs
@@ -3,10 +3,10 @@ use ratatui::buffer::Buffer;
 use ratatui::layout::Rect;
 use ratatui::widgets::WidgetRef;
 
+use super::list_selection_view::GenericDisplayRow;
+use super::list_selection_view::ScrollState;
+use super::list_selection_view::render_rows;
 use super::popup_consts::MAX_POPUP_ROWS;
-use crate::selection_menu::GenericDisplayRow;
-use crate::selection_menu::ScrollState;
-use crate::selection_menu::render_rows;
 
 /// Visual state for the file-search popup.
 pub(crate) struct FileSearchPopup {
