--- codex-rs/tui/src/selection_menu.rs
@@ -1,235 +1,0 @@
-use ratatui::buffer::Buffer;
-use ratatui::layout::Rect;
-use ratatui::prelude::Constraint;
-use ratatui::style::Color;
-use ratatui::style::Modifier;
-use ratatui::style::Style;
-use ratatui::style::Stylize;
-use ratatui::text::Line;
-use ratatui::text::Span;
-use ratatui::widgets::Block;
-use ratatui::widgets::BorderType;
-use ratatui::widgets::Borders;
-use ratatui::widgets::Cell;
-use ratatui::widgets::Row;
-use ratatui::widgets::Table;
-use ratatui::widgets::Widget;
-
-/// Generic scroll/selection state for a vertical list menu.
-///
-/// Encapsulates the common behavior of a selectable list that supports:
-/// - Optional selection (None when list is empty)
-/// - Wrap-around navigation on Up/Down
-/// - Maintaining a scroll window (`scroll_top`) so the selected row stays visible
-#[derive(Debug, Default, Clone, Copy)]
-pub(crate) struct ScrollState {
-    pub selected_idx: Option<usize>,
-    pub scroll_top: usize,
-}
-
-impl ScrollState {
-    pub fn new() -> Self {
-        Self {
-            selected_idx: None,
-            scroll_top: 0,
-        }
-    }
-
-    /// Reset selection and scroll.
-    pub fn reset(&mut self) {
-        self.selected_idx = None;
-        self.scroll_top = 0;
-    }
-
-    /// Clamp selection to be within the [0, len-1] range, or None when empty.
-    pub fn clamp_selection(&mut self, len: usize) {
-        self.selected_idx = match len {
-            0 => None,
-            _ => Some(self.selected_idx.unwrap_or(0).min(len.saturating_sub(1))),
-        };
-        if len == 0 {
-            self.scroll_top = 0;
-        }
-    }
-
-    /// Move selection up by one, wrapping to the bottom when necessary.
-    pub fn move_up_wrap(&mut self, len: usize) {
-        if len == 0 {
-            self.selected_idx = None;
-            self.scroll_top = 0;
-            return;
-        }
-        self.selected_idx = Some(match self.selected_idx {
-            Some(idx) if idx > 0 => idx - 1,
-            Some(_) => len - 1,
-            None => 0,
-        });
-    }
-
-    /// Move selection down by one, wrapping to the top when necessary.
-    pub fn move_down_wrap(&mut self, len: usize) {
-        if len == 0 {
-            self.selected_idx = None;
-            self.scroll_top = 0;
-            return;
-        }
-        self.selected_idx = Some(match self.selected_idx {
-            Some(idx) if idx + 1 < len => idx + 1,
-            _ => 0,
-        });
-    }
-
-    /// Adjust `scroll_top` so that the current `selected_idx` is visible within
-    /// the window of `visible_rows`.
-    pub fn ensure_visible(&mut self, len: usize, visible_rows: usize) {
-        if len == 0 || visible_rows == 0 {
-            self.scroll_top = 0;
-            return;
-        }
-        if let Some(sel) = self.selected_idx {
-            if sel < self.scroll_top {
-                self.scroll_top = sel;
-            } else {
-                let bottom = self.scroll_top + visible_rows - 1;
-                if sel > bottom {
-                    self.scroll_top = sel + 1 - visible_rows;
-                }
-            }
-        } else {
-            self.scroll_top = 0;
-        }
-    }
-}
-
-/// A generic representation of a display row for selection popups and menus.
-pub(crate) struct GenericDisplayRow {
-    pub name: String,
-    pub match_indices: Option<Vec<usize>>, // indices to bold (char positions)
-    pub is_current: bool,
-    pub description: Option<String>, // optional grey text after the name
-}
-
-impl GenericDisplayRow {}
-
-/// Render a list of rows using the provided ScrollState, with shared styling
-/// and behavior for selection popups and menus.
-pub(crate) fn render_rows(
-    area: Rect,
-    buf: &mut Buffer,
-    rows_all: &[GenericDisplayRow],
-    state: &ScrollState,
-    max_results: usize,
-    _dim_non_selected: bool,
-    empty_message: &str,
-) {
-    let mut rows: Vec<Row> = Vec::new();
-    if rows_all.is_empty() {
-        rows.push(Row::new(vec![Cell::from(Line::from(
-            empty_message.dim().italic(),
-        ))]));
-    } else {
-        let max_rows_from_area = area.height as usize;
-        let visible_rows = max_results
-            .min(rows_all.len())
-            .min(max_rows_from_area.max(1));
-
-        // Compute starting index based on scroll state and selection.
-        let mut start_idx = state.scroll_top.min(rows_all.len().saturating_sub(1));
-        if let Some(sel) = state.selected_idx {
-            if sel < start_idx {
-                start_idx = sel;
-            } else if visible_rows > 0 {
-                let bottom = start_idx + visible_rows - 1;
-                if sel > bottom {
-                    start_idx = sel + 1 - visible_rows;
-                }
-            }
-        }
-
-        for (i, row) in rows_all
-            .iter()
-            .enumerate()
-            .skip(start_idx)
-            .take(visible_rows)
-        {
-            let GenericDisplayRow {
-                name,
-                match_indices,
-                is_current: _is_current,
-                description,
-            } = row;
-
-            // Highlight fuzzy indices when present.
-            let mut spans: Vec<Span> = Vec::with_capacity(name.len());
-            if let Some(idxs) = match_indices.as_ref() {
-                let mut idx_iter = idxs.iter().peekable();
-                for (char_idx, ch) in name.chars().enumerate() {
-                    if idx_iter.peek().is_some_and(|next| **next == char_idx) {
-                        idx_iter.next();
-                        spans.push(ch.to_string().bold());
-                    } else {
-                        spans.push(ch.to_string().into());
-                    }
-                }
-            } else {
-                spans.push(name.clone().into());
-            }
-
-            if let Some(desc) = description.as_ref() {
-                spans.push("  ".into());
-                spans.push(desc.clone().dim());
-            }
-
-            let mut cell = Cell::from(Line::from(spans));
-            if Some(i) == state.selected_idx {
-                cell = cell.style(
-                    Style::default()
-                        .fg(Color::Cyan)
-                        .add_modifier(Modifier::BOLD),
-                );
-            }
-            rows.push(Row::new(vec![cell]));
-        }
-    }
-
-    let table = Table::new(rows, vec![Constraint::Percentage(100)])
-        .block(
-            Block::default()
-                .borders(Borders::LEFT)
-                .border_type(BorderType::QuadrantOutside)
-                .border_style(Style::default().add_modifier(Modifier::DIM)),
-        )
-        .widths([Constraint::Percentage(100)]);
-
-    table.render(area, buf);
-}
-
-#[cfg(test)]
-mod tests {
-    use super::ScrollState;
-
-    #[test]
-    fn wrap_navigation_and_visibility() {
-        let mut s = ScrollState::new();
-        let len = 10;
-        let vis = 5;
-
-        s.clamp_selection(len);
-        assert_eq!(s.selected_idx, Some(0));
-        s.ensure_visible(len, vis);
-        assert_eq!(s.scroll_top, 0);
-
-        s.move_up_wrap(len);
-        s.ensure_visible(len, vis);
-        assert_eq!(s.selected_idx, Some(len - 1));
-        match s.selected_idx {
-            Some(sel) => assert!(s.scroll_top <= sel),
-            None => panic!("expected Some(selected_idx) after wrap"),
-        }
-
-        s.move_down_wrap(len);
-        s.ensure_visible(len, vis);
-        assert_eq!(s.selected_idx, Some(0));
-        assert_eq!(s.scroll_top, 0);
-    }
-}
